% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{tissues}
\alias{tissues}
\title{Human Tissue Experiment (Wang et al.)}
\format{List with 3 elements:
\describe{
  \item{Expression}{Integer matrix: Features by samples matrix with integers representing raw counts}
  \item{Design}{Dataframe: Study design}
  \item{Annotation}{Dataframe: Additional info about features}
  The three objects are already in the same order, but can be matched using row/column names.
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed?term=18978772}
}
\usage{
tissues
}
\description{
RNA-Seq data originally from RECount. For more information, see reference from URL.
}
\keyword{datasets}

